Improvements to JS1 week1 homework
* Tests show what was expected and what was actually returned. This both
  helps the students to understand what was wrong, and is more like what
  they'll encounter in real life from test suites.
* Fix up partial rename from getRemainder to getTotal.
* Fix up real-world logic of currency conversion (a % fee should _cost_
  you more, not _give_ you more!)
* Magic 8 ball doesn't use a global variable, and explains slightly more
  what checkAnswer should actually do.
* Magic 8 ball tests to make sure more than one value is actually
 returned.

Note that this introduces use of `require` to import `util` - this is a
built-in module in Node, so doesn't add any extra requirements to how
the tests are run, but does introduce an element of syntax we haven't
taught the students. This is only in the test machinery itself, and is
fairly cosmetic, so hopefully won't confuse folks.